+++ packages/sif-common-core-ds/src/index.ts
@@ -1,9 +1,7 @@
-export { default as FormikAttachmentForm } from './components/formik-attachment-form/FormikAttachmentForm';
 export { default as FormikFileUpload } from './components/formik-file-upload/FormikFileUpload';
 export { default as VedleggList } from './components/vedlegg-list/VedleggList';
 export { default as VedleggSummaryList } from './components/vedlegg-summary-list/VedleggSummaryList';
 export { getVedleggValidator } from './components/formik-file-upload/getVedleggValidator';
-export * from './components/formik-attachment-form/getAttachmentsValidator';
 export * from './hooks';
 export * from './types';
 export * from './utils/apiUtils';