+++ packages/sif-common-core-ds/src/types/PersistedFile.ts
@@ -1,0 +1,7 @@
+export interface PersistedFile {
+    isPersistedFile: boolean;
+    name: string;
+    lastModified: number;
+    size: number;
+    type: string;
+}