+++ packages/sif-common-core-ds/storybook/stories/components/FormikFileUpload.stories.tsx
@@ -1,7 +1,7 @@
 import { Box, Heading, VStack } from '@navikt/ds-react';
 import { Meta, StoryObj } from '@storybook/react';
 import * as React from 'react';
-import { getAttachmentsValidator } from '@navikt/sif-common-core-ds';
+import { getVedleggValidator } from '@navikt/sif-common-core-ds';
 import { useFormikContext } from 'formik';
 import { http, HttpResponse } from 'msw';
 import { v4 as uuidv4 } from 'uuid';
@@ -49,10 +49,10 @@ const Example = () => {
                 label="Last opp dokumenter"
                 useDefaultDescription={true}
                 initialFiles={values.vedlegg}
-                validate={getAttachmentsValidator(
+                validate={getVedleggValidator(
                     {
                         errors: {
-                            noAttachmentsUploaded: {
+                            noVedleggUploaded: {
                                 keyPrefix: 'validation.arbeidsforhold.utbetalingsårsak.vedlegg',
                                 keepKeyUnaltered: true,
                                 values: { arbeidsgivernavn: 'abc' },